Rename package: worldgen → auroch_syna
