+++ src/auroch_syna/worldgen/pano_seg.py
