+++ src/auroch_syna/worldgen/utils/equirectangular.py
