+++ src/auroch_syna/worldgen/utils/splat_utils.py
